Show current and accumulate damage for each player
Each shot sent between players are confirmed via interactions topic, the
frontend captures the interactions message and present them spliting
between Elixir and Scala damage.

The idea is to change on-the-fly the shots interval or even pause one of
the players and in the prime case reset the Kafka offset on-the-fly

In the end under a high pressure which player will be more resilient to
hold the load (message lagging) and catch up properly.